fix : улучшения retry функции
